# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Image Processing Assignment API",
    "description": "Asynchronous image processing job API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    }
  ],
  "tags": [
    {
      "name": "Jobs",
      "description": "Asynchronous image processing jobs"
    }
  ],
  "paths": {
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List jobs",
        "operationId": "getJobs",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page number",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size. Maximum is 100.",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid paging parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidSize": {
                    "value": {
                      "code": "BAD_REQUEST",
                      "message": "getJobs.size: size must be at most 100"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create image processing job",
        "description": "Creates a Job and returns immediately. The image is processed asynchronously by Mock Worker.",
        "operationId": "createJob",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Idempotency key for duplicate request handling",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              },
              "examples": {
                "createJob": {
                  "value": {
                    "imageUrl": "https://example.com/image.png"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidRequest": {
                    "value": {
                      "code": "BAD_REQUEST",
                      "message": "imageUrl: imageUrl must be a valid https URL"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "idempotencyConflict": {
                    "value": {
                      "code": "IDEMPOTENCY_CONFLICT",
                      "message": "Idempotency key conflict. idempotencyKey=sample-key-1"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get job",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notFound": {
                    "value": {
                      "code": "NOT_FOUND",
                      "message": "Job not found. jobId=00000000-0000-0000-0000-000000000000"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateJobRequest": {
        "required": [
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "imageUrl": {
            "maxLength": 2048,
            "minLength": 1,
            "type": "string",
            "format": "uri",
            "example": "https://example.com/image.png"
          }
        }
      },
      "JobResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/image.png"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "result": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "JobPageResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PROCESSING",
          "COMPLETED",
          "FAILED"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "imageUrl: imageUrl must be a valid https URL"
          }
        }
      }
    }
  }
}